+++ generator/src/main/java/ai/care/arc/generator/codegen/spec/FieldSpecGenSetter.java
@@ -1,4 +1,4 @@
-package ai.care.arc.generator.codegen;
+package ai.care.arc.generator.codegen.spec;
 
 import com.squareup.javapoet.FieldSpec;
 import com.squareup.javapoet.MethodSpec;
