+++ generator/src/test/java/ai/care/arc/generator/codegen/spec/FieldSpecGenSetterTest.java
@@ -1,5 +1,6 @@
-package ai.care.arc.generator.codegen;
+package ai.care.arc.generator.codegen.spec;
 
+import ai.care.arc.generator.codegen.spec.FieldSpecGenSetter;
 import com.squareup.javapoet.FieldSpec;
 import com.squareup.javapoet.MethodSpec;
 import org.junit.Test;
